requirements.txt
django>=4.2
django-role-permissions>=3.2
djangorestframework>=3.15
python-dotenv>=1.0
django-ratelimit>=4.1
